# [diff] Introduce a concurrent reader to optimize unpark