The concurrent reader will start a goroutine in the background to read data into the buffer, then send buffer to chan. the foreground Reader will receive the buffer from chan and return it to the caller.
This makes decompression and untar process in parallel. 

### Optimization results:
Use `ctr I pull busybox:latest` and collect the output `done: xxx ms`.

| image | image size | now | conReader | % improvement |
| --- | --- | --- | --- | --- |
| bysubox | 4.86MB | 52.458629ms| 48.593998ms |  7.36702% |
| coredns | 53.6MB | 246.313598ms | 227.530718ms | 7.6256% |
| nginx | 142MB | 983.486548ms | 876.264884ms | 10.9022% |
| openjdk | 470MB | 2.254758463s | 1.997340339s | 11.4167% |
| node | 1GB | 7.076312855s | 5.490200976s | 22.4144% |

Testing CPU/Memory usage with `node` images. 
Details for `node` images.

```
KEY                                                                     SIZE      INODES 
sha256:6d7140527100b8a0b62be606f36279d068bc6a0fa52a415f02d8dd8930e4a226 20.0 KiB  5      
sha256:4017cd6327da64d694e907023bd3b27cf7073ced8456e88014989b41cdacba41 7.3 MiB   32     
sha256:37cd3570fe018fbedb5359582a8d749c9ba8a3806519649433cef62df9656b3a 165.0 MiB 5545   
sha256:6c57c349090475e4c491e2c774e2cb25cd949db386d5bc0ac605908b65d10584 396.0 KiB 24     
sha256:230c3b2acf5c7ce5ccbc11c5292092a5af0ce5b1e366228eee031eb9c84ed0dc 532.3 MiB 12785  
sha256:34349aaf2e7f75c266222ea09df6e2bef39fea73640d191dfc5d271d390103fc 159.9 MiB 6197   
sha256:f51029f953be13d297d2bc2dd7c56c0a84c77cd7bd6672837a01c4571c7cf5fa 19.4 MiB  502    
sha256:2fab58d5cf486ed4ca265634fde3c8b4a046ade908d9b1d5405d8dfce6d12413 12.3 MiB  958    
sha256:ae56c0c5405b3d38a59b5ec5dfe5db36e8b6ff6ca8595851f903cde13e709282 134.7 MiB 7248 
```
### CPU usage

* now
![image](https://github.com/containerd/containerd/assets/8768126/ee2c8440-ebc9-424e-a5bc-22fd32d6e88b)
* conReader
![image](https://github.com/containerd/containerd/assets/8768126/16cae61e-a02f-446d-ac8a-b8b19d96f319)

### Memory usage
Testing with `node` images.
* now
![image](https://github.com/containerd/containerd/assets/8768126/7608d605-1a2d-413c-9d9f-8959cc1f7f89)
![image](https://github.com/containerd/containerd/assets/8768126/02130cea-b04f-4ab3-89ee-6782422ec29d)

* conReader
![image](https://github.com/containerd/containerd/assets/8768126/40b3878a-15e9-4c9f-951f-c241a4f8c7e7)
![image](https://github.com/containerd/containerd/assets/8768126/d83de1b5-d682-4a0e-bd54-c3a27d037953)