@fuweid 2023-05-19
It seems like that bottleneck is caused by digest cal. I'm just worry about the memory usage. Maybe we need to verify the usage with 64/MiB128MiB/256MiB/512MiB/1GiB dataset.