@wllenyj 2023-05-19
> This is the table above, but with some percentages to make the improvement easier to see.
> 
> image	now	conReader	% improvement
> busybox	52.458629ms	48.593998ms	7.36702%
> coredns	246.313598ms	227.530718ms	7.6256%
> nginx	983.486548ms	876.264884ms	10.9022%
> openjdk	2.254758463s	1.997340339s	11.4167%
> node	7.076312855s	5.490200976s	22.4144%

Thank you very much. Updated