@ruiwen-zhao 2023-05-19
> It seems like that bottleneck is caused by digest cal. I'm just worry about the memory usage. Maybe we need to verify the usage with 64/MiB128MiB/256MiB/512MiB/1GiB dataset.

This change improves things by making unpacking and decompression parallel, so the bottleneck is not necessarily the digest calculation right? 

Have we looked at disk IO usage before and after this change, since this change parallels io operations as well ?