@fuweid 2023-05-19
@ruiwen-zhao I was running into the issue about digest calculations and the issue #8506 also mentions it. I don't have strong opinions on this, since 7 -> 5 seconds is not bad. If the memory can be lowered, it will be great.