@cpuguy83 2023-05-19
Have we looked at using a bufio Reader here instead?